+++ lib/jarbler/JarMain.java
@@ -153,9 +153,9 @@ class JarMain {
             debug("JRuby set property 'jruby.gem.home' to '" + full_gem_home + "'");
             System.setProperty("jruby.gem.home", full_gem_home);
 
-            String full_gem_path = full_gem_home +  File.pathSeparator + full_gem_home + File.separator + "bundler" ;
-            debug("JRuby set property 'jruby.gem.path' to '" + full_gem_path + "'");
-            System.setProperty("jruby.gem.path", full_gem_path);
+            // String full_gem_path = full_gem_home +  File.pathSeparator + full_gem_home + File.separator + "bundler" ;
+            // debug("JRuby set property 'jruby.gem.path' to '" + full_gem_path + "'");
+            // System.setProperty("jruby.gem.path", full_gem_path);
 
             debug("JRuby program starts with the following arguments: ");
             for (String arg : mainArgs) {
@@ -245,7 +245,7 @@ class JarMain {
 
     private static void debug(String msg) {
         if (System.getenv("DEBUG") != null) {
-            System.out.println(msg);
+            System.err.println(msg);
         }
     }