+++ test/test_helper.rb
@@ -4,8 +4,10 @@ require 'open3'
 class Minitest::Test
 
   def setup
-    puts "\n\n-------------------------------------------------------------------"
-    log "##### Starting test #{self.class.name}::#{self.name}"
+    start_msg = "Starting test #{self.class.name}::#{self.name}"
+    puts "\n\n#{'-' * (start_msg.length + 20)}"
+    log start_msg
+    puts '-' * (start_msg.length + 20)
     debug "Gem.paths.path in setup: #{Gem.paths.path}"
     debug "GEM_HOME in setup: #{ENV['GEM_HOME']}" if ENV['GEM_HOME']
     debug "GEM_PATH in setup: #{ENV['GEM_PATH']}" if ENV['GEM_PATH']
@@ -14,7 +16,10 @@ class Minitest::Test
   end
 
   def teardown
-    log "##### End test #{self.class.name}::#{self.name}\n\n"
+    end_msg = "End of test #{self.class.name}::#{self.name}"
+    puts "#{'-' * (end_msg.length + 20)}"
+    log end_msg
+    puts '-' * (end_msg.length + 20)
     super
   end
 
@@ -26,11 +31,11 @@ class Minitest::Test
     log(msg) if ENV['DEBUG']
   end
 
-  def exec_and_log(command)
+  def exec_and_log(command, env: {})
     log("Execute by Open3.capture3: #{command}")
-    stdout, stderr, status = Open3.capture3(command)
+    stdout, stderr, status = Open3.capture3(env, command)
     log("Command '#{command}'Executed with  Open3.capture3: status = #{status}\nstdout:\n#{stdout}\n\nstderr:\n#{stderr}\n")
-    assert status.success?, "Response status should be success but is:\n#{stdout}\nstderr:\n#{stderr}\nstatus: #{status}"
+    assert status.success?, "Response status should be success but is '#{status}':\n#{stdout}\nstderr:\n#{stderr}"
     return stdout, stderr, status
   end
 end